add an amplitude effect on the yellow LEDs
We want a simple value for amplitude, and the best way to test is with
something visual. This implementation uses a 16 bit value for the LEDs
translated from the 24 bit data stream.

There is no amplitude adjustment internally to keep the values within a
proper range for the LEDs. This means that if the volume of the input
gets too loud it will clip off the top of the LEDs and they will all be
lit all the time. To solve this, turn the volume down. More work could
be done on this if we had time.

Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>