ops, forgot a couple of zeros
The audio amplitude is based on a 23 bit number (24 bit signed). I make
the range 130000-1430000. This was a trial and error process of where
most music tends to look best. To be fair, I could have adjusted
the actual amplitude into a range that acurately reflected it, but we
want the demo to look good. We can worry about acuracty later.

Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>